requirements.txt
environs~=9.0
gino==1.0.1
sqlalchemy==1.3.24
fastapi-cache2[redis]